+++ src/main/resources/templates/blocks/footer.html
@@ -1,9 +1,6 @@
 <div th:fragment="footer">
   <footer class="text-muted py-5">
     <div class="container">
-      <p class="float-end mb-1">
-        <a href="#">Back to top</a>
-      </p>
       <p class="mb-0">Вас что-то волнует? Расскажите об этом в посте!</p>
     </div>
   </footer>
