+++ src/main/resources/templates/postAdd.html
@@ -11,6 +11,7 @@
 <div class="container mt-5 mb-5">
   <h1>Добавление статьи</h1>
   <form action="/blog/addPost" method="post">
+    <input type="hidden" name="_csrf" value="{{_csrf.token}}" />
     <input type="text" name="subjectPost" placeholder="Введите название статьи" class=" form-control"  th:value="${post.subjectPost}"><br>
     <input type="text" name="anonsPost" placeholder="Введите анонс статьи" class=" form-control" th:value="${post.anonsPost}"><br>
