+++ src/main/resources/templates/profileDetails.html
@@ -1,51 +1,19 @@
 <!DOCTYPE HTML>
 <html xmlns:th="http://www.thymeleaf.org" xmlns="http://www.w3.org/1999/html">
-<link rel="stylesheet" href="https://bootstraptema.ru/plugins/2015/bootstrap3/bootstrap.min.css" />
-<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css" />
-<br><br><br>
+<head>
+    <title th:text="${title}" />
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css">
+    <link rel="stylesheet" href="https://bootstraptema.ru/plugins/2015/bootstrap3/bootstrap.min.css" />
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css" />
+</head>
+<body>
 
-<style>
-body{background:url(https://bootstraptema.ru/images/bg/bg-1.png)}
+<header th:insert="/blocks/header :: header"></header>
 
-#main {
- background-color: #f2f2f2;
- padding: 20px;
--webkit-border-radius: 4px;
- -moz-border-radius: 4px;
- -ms-border-radius: 4px;
- -o-border-radius: 4px;
- border-radius: 4px;
- border-bottom: 4px solid #ddd;
-}
-#real-estates-detail #author img {
- -webkit-border-radius: 100%;
- -moz-border-radius: 100%;
- -ms-border-radius: 100%;
- -o-border-radius: 100%;
- border-radius: 100%;
- border: 5px solid #ecf0f1;
- margin-bottom: 10px;
-}
-#real-estates-detail .sosmed-author i.fa {
- width: 30px;
- height: 30px;
- border: 2px solid #bdc3c7;
- color: #bdc3c7;
- padding-top: 6px;
- margin-top: 10px;
-}
-.panel-default .panel-heading {
- background-color: #fff;
-}
-#real-estates-detail .slides li img {
- height: 450px;
-}
-</style>
-
+<br>
 <div class="container">
     <div id="main">
-
-
         <div class="row" id="real-estates-detail">
             <div class="col-lg-4 col-md-4 col-xs-12">
                 <div class="panel panel-default">
@@ -73,13 +41,12 @@ body{background:url(https://bootstraptema.ru/images/bg/bg-1.png)}
                                     <a style="float: right" class="me-3 py-2 text-dark text-decoration-none" th:href="@{'/blog/profile/' + ${author.id}+'/edit'}">Редактировать страницу</ya-tr-span></a>
                                 <table class="table table-th-block">
                                     <tbody>
-                                    <tr><td class="active">Зарегистрирован:</td><td>12-06-2016</td></tr>
-                                    <tr><td class="active">Последняя активность:</td><td>12-06-2016 / 09:11</td></tr>
                                     <tr><td class="active">Пол:</td><td>...</td></tr>
-                                    <tr><td class="active">Полных лет:</td><td>43</td></tr>
-                                     </tbody>
+                                    <tr><td class="active">Полных лет:</td><td> <p th:if="${age} != 0" th:text="${age}"></p></td></tr>
+                                    <tr><td class="active">Зарегистрирован:</td><td><p th:text="${author.dateCreate}"></p></td></tr>
+                                    </tbody>
                                 </table>
-                                <div class="container mt-5">
+                                <div class="container-fluid">
                                     <h1>Последние публикации</h1>
                                     <div th:each = "el : ${posts}" class="alert alert-info mt-2">
                                         <div style="float: right" th:text="${el.Author.nickname}">Автор</div>
@@ -99,4 +66,7 @@ body{background:url(https://bootstraptema.ru/images/bg/bg-1.png)}
     </div>
 
 </div><!-- /.main -->
+<div th:insert="/blocks/footer :: footer"></div>
+
+</body>
 </html><!-- /.container -->
